refactor: Delete legacy breadcrumb component
